buildroot: Compress images with zstd
My benchmarks show that zstd has the second best decompression
ratio/speed, whereas the best decompression ratio algorithm is
not the same as the best decompression speed algorithm. Compression
speed is not awful either but that is not as much of a concern because
compression only happens during image build time, where it is a very
small fraction of the overall build process.

Check that the zstd tool is installed and bail out if it is not.

Link: https://github.com/ClangBuiltLinux/continuous-integration/issues/184#issuecomment-604221248
Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>