boot-qemu.sh: Initial version
To quote Nick:

  "Orthogonally, I've been finding it frustrating to:

  1. jump on a new machine without shell history, and try to recreate
     our QEMU commands from driver.sh. Even sending commands to newbies
     is a PITA.
  2. boot more than just our scripts that immediately power off a
     machine."

driver.sh is a very basic kernel build script, it is not good for
extensive testing because it only builds a defconfig kernel with no
modules. However, the boot testing that it provides is invaluable
because we have figured out all of the necessary QEMU flags.

Enter boot-qemu.sh, a split out version of driver.sh's booting
framework. There are two required arguments: the architecture being
booted (which expects the same value as driver.sh does) and a path to
the kernel's object folder (either the kernel source itself if an
in-tree build or the path to O= if an out-of-tree build).

By default, the rootfs will launch /init, which will execute our init
script that just prints the kernel version and shuts the machine down.
The script has an option to dump into an interactive shell (busybox's
shell) so that more testing can be done.

Run:

  $ ./boot-qemu.sh -h

for more information.

Suggested-by: Nick Desaulniers <ndesaulniers@google.com>
Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>