boot-qemu.sh: Hoist '-nodefaults' from ppc64le arguments to main arguments
This prevents QEMU from creating devices underneath us, which can change
across QEMU versions (see de875be20ca66). To prevent this in the future,
add '-nodefaults' for all invocations of QEMU, not just powerpc64le.

Tested with QEMU 3.1.0, 4.2.0, and 5.0.0-rc2.

NOTE: This is the only architecture that we actually set devices for so
this should be a no-op for everything else but we might add other
devices in the future.

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>