boot-qemu.sh: Combine x86 and x86_64 for shared KVM handling
The x86 section unconditionally adds -enable-kvm to the QEMU flags. This
won't work for Travis because it does not have KVM support. The x86_64
section handles this already and the only difference between the x86 and
x86_64 sections is the QEMU binary name. Combine the sections and handle
that so that there is less duplication and everything works properly.

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>