run timeout in the foreground so it can be killed
When a kernel is hung, we currently have to wait the timeout to exit,
(or ctrl+z, pkill timeout).  For complicated reasons, running timeout
from a shell script won't forward signals to timeout.

Adding --foreground is the simplest option for this to work, via:
https://unix.stackexchange.com/a/233685/388381

See other answers in that thread for *why* this is the case.