boot-qemu.sh: Break apart decomp_rootfs
There is no point in decompressing the rootfs before we make sure that
the kernel image and dtb can be found. Break this function apart because
we need to know the image location before ${ARCH} is reassigned but we
can decompress the image once we know we are actually going to call the
QEMU binary.

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>