boot-utils: Add support for Debian images
There are times where there is a need to have a fuller environment for
testing, rather than the simple Busybox based rootfs that is normally
used.

Introduce a Debian build script, which uses debootstrap and
qemu-user-static to generate a bootable rootfs image.

boot-qemu.sh is modified accordingly to support this image if the user
wants it.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>