boot-qemu.sh: Use Nehalem for non-KVM boot on x86_64
We are going to update our x86_64 rootfs image to be able to be built
with a prebuilt toolchain to speed up builds of the rootfs images, which
will require at least SSE 4.2 support. This CPU is the microarchitecture
of the first Core i7 processor, which will be the target for Buildroot.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>