buildroot/mips*.config: Use Bootlin's prebuilt toolchain
This saves cycles while rebuilding the images.

We need to switch off software floating point emulation so that the
prebuilt toolchain can be selected. This should not impact anything as
all of the kernels that we build have CONFIG_MIPS_FP_SUPPORT, which
emulates an FPU if there is not one present.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>