buildroot/ppc*.config: Use Bootlin's prebuilt toolchain
This saves cycles while rebuilding the images.

For 32-bit, we need to switch to the 440fp variant so that the prebuilt
toolchain can be selected. This does not impact anything because the
kernel builds FP emulation for targets that do not have an FPU (which it
seems like bamboo does from my research anyways).

For 64-bit, we need to switch to the power8 variant so that the prebuilt
toolchain can be selected. We can boot QEMU with a power8 CPU so this
does not affect anything either.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>