buildroot/riscv.config: Use Bootlin's prebuilt toolchain
This saves cycles while rebuilding the images.

We need to switch to the LP64 ABI so that the prebuilt toolchain can be
selected. This is a soft float ABI, which is not the biggest deal
because this rootfs will always be run under emulation so performance is
not the biggest deal.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>