boot-qemu.sh: Add support for earlycon
There is currently a crash on -next for RISC-V that is not visible
without this. For all ARM architectures supported with this script, just
"earlycon" works because the device trees have support for the console
and baud rates. RISC-V is the same way. x86 needs the type of I/O and
address. ttyS0 is 0x3f8 (arch/x86/boot/early_serial_console.c) and we
do regular I/O. The baud rate works with the default so it is omitted.

MIPS already has earlycon enabled.

I cannot find any research on earlycon on PowerPC or s390 so they are
omitted.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>