debian/build.sh: qemu-debootstrap -> debootstrap
qemu-debootstrap has been deprecated in favor of just running
debootstrap due to the use of the "F" flag with binfmt_misc.

https://salsa.debian.org/qemu-team/qemu/-/commit/9899b1723119e57354c5d8e6f68f15ed762bd570

This makes it easier for other distributions to run the script because
debootstrap is usually available from the package manager but
qemu-debootstrap is a part of Debian's qemu-user-static package.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>