boot-qemu.sh: Add 'arm' as an alias for 'arm32_v7'
This is an opinionated change but one I think most people will agree
with. I boot test multi_v7_defconfig variants more than either of the
other 32-bit ARM configurations we test and remembering to put the
'32_v7' part is rather annoying. Add 'arm' as a valid '-a' value so that
it works the same as 'arm32_v7'.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>