boot-qemu.sh: Clamp '-smp' based on CONFIG_NR_CPUS when available
As noted on the review for #17, when booting an i386_defconfig kernel,
there is a lot of spam in the boot logs because CONFIG_NR_CPUS is set to
8:

[    0.020644] APIC: NR_CPUS/possible_cpus limit of 8 reached. Processor 8/0x8 ignored.
[    0.021302] APIC: NR_CPUS/possible_cpus limit of 8 reached. Processor 9/0x9 ignored.
[    0.021964] APIC: NR_CPUS/possible_cpus limit of 8 reached. Processor 10/0xa ignored.
[    0.022642] APIC: NR_CPUS/possible_cpus limit of 8 reached. Processor 11/0xb ignored.

With large machines, this can be quite noisy, as there is one print out
for every CPU above 8.

Refactor the logic for finding a sensible value based on CONFIG_NR_CPUS
if a config can be found. If no config is available, default to 8, which
is the default value for i386_defconfig.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>