boot-qemu.sh: Add a function to get the kernel version as a number
Similar to the Linux kernel's representation of compiler versions, add
the ability to get the version of the kernel being booted as a six or
seven digit number to make decisions based on that number.

To do this, we need a decompressed kernel image, so that we can use
"strings" + "grep". The arguments to the function are the command to
dump the decompressed kernel image to stdout. For example, if the kernel
image is Image.gz, the command would be "gzip -c -d ...".

Signed-off-by: Nathan Chancellor <nathan@kernel.org>