buildroot: Build an ext4 image for x86_64
I have not found a clean way to boot User Mode Linux with our existing
cpio rootfs images but it does work with a standard ext4 image. Tell
buildroot to build an ext4 image as part of the build process for
x86_64, which can be used for UML. This extra image only adds 1.61MB to
the repo when compressed.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>