boot-qemu.sh: Use different CPU for arm64 with new QEMU + old kernel
After FEAT_LPA was implemented in QEMU's TCG, kernels without
commit 787fd1d019b2 ("arm64: limit PA size to supported range") fail to
boot; this change first showed up in 4.16.

We already have a workaround for FEAT_LPA2 not working with kernels
older than 5.12, which is known. Reporting this to QEMU upstream
resulted in the recommendation to use one of the ARMv8.0 CPUs for these
older kernels. Use 'cortex-a72' for booting anything older than 4.16
with QEMU 6.2.50 (master) and newer.

Link: https://gitlab.com/qemu-project/qemu/-/issues/964
Signed-off-by: Nathan Chancellor <nathan@kernel.org>