boot-utils: Remove skiboot
We had initially shipped this to workaround issues with Debian's version
of skiboot back in 2019:

https://github.com/ClangBuiltLinux/continuous-integration/commit/ea12bb8463bf331b1aa9202e729080d266b7d26f

Since then, it has seen a couple of updates, the last of which was done
on March 25th, 2020 (a little under two and a half years ago):

https://github.com/ClangBuiltLinux/continuous-integration/commit/989c7b1d08b8e12b506198e47e862c180fac833c
https://github.com/ClangBuiltLinux/boot-utils/commit/51bc49ee147ad6903c234d8360a9fbbb4d7e1dd8

The most recent skiboot release is 7.0, which was released on October
25th, 2021 (almost a year old).

Instead of updating skiboot to the latest release, just remove this
vendored version altogether and rely on the distribution's version of
skiboot. While this may impact users of older distributions, it lowers
the maintainence overhead for boot-utils and makes it less likely that
we run into a problem with newer QEMU but older skiboot.

Debian stable currently ships skiboot 6.4 in the qemu-system-data
package and Arch Linux (used by continuous-integration2) currently ships
skiboot 7.0 so this should have little impact to users of up to date
distributions.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>