boot-qemu.py: Add support for booting arm64 and x86_64 guests under UEFI
Booting under UEFI can expose potential issues that will impact real
hardware, like CFI failures. Add support for booting under UEFI with
arm64 and x86_64 by searching for firmare on the user's system, copying
and modifying it as necessary, then passing it along to QEMU.

This is deliberately a separate option that is off by default, as I am
not sure we can turn this on by default without regressing older trees
or toolchains. It has only been tested with tip of tree Linux and LLVM.

Closes: https://github.com/ClangBuiltLinux/boot-utils/issues/14
Signed-off-by: Nathan Chancellor <nathan@kernel.org>