boot-qemu.py: Break launch_qemu() into smaller functions
launch_qemu() has gotten quite unweildy in terms of indentation level
and readability. Fix it by breaking it a part into several different
functions, each of which has full documentation behind it to understand
what they are doing and why they exist.

Take the opportunity to optimize launch_qemu_gdb() a bit by pulling as
many unnecessary changes out of the while loop as possible.

There are a couple of minor changes around spacing in this, which came
about due to changes to the order in which strings are printed.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>