boot-qemu.py: Simplify can_use_kvm()
The existing check for '/dev/kvm' is not strong enough, as it is
possible that '/dev/kvm' exists but the user does not have sufficient
permissions to access it, which will result in QEMU immediately refusing
to start. This can be quite frustrating when doing a sequence of builds
non-interactively, as the results will be tainted due to the initial
startup failure.

Switch to os.access(), which allows us to make sure '/dev/kvm' is
writeable with the current permissions.

At the same time, clean up this function. If '/dev/kvm' is accessible,
it implies hardware virtualization support, so we can ditch the
'/proc/cpuinfo' check for x86. Refactor the 32-bit EL1 check for aarch64
to its own function, which allows us to compact the aarch64 section
signficantly.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>