boot-qemu.py: Use more specific exception class
New pylint warns:

  boot-qemu.py:401:12: W0719: Raising too general exception: Exception (broad-exception-raised)
  boot-qemu.py:431:16: W0719: Raising too general exception: Exception (broad-exception-raised)

The type of exception really does not matter, as we do not expect this script
to be called, so the exception will never be caught. However, it is easy to
silence this by switching to a more specific exception. FileNotFoundError is
descriptive so use this exception to silence the warning.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>