boot-qemu.py: Add ppc64 big endian ELFv2 file strings to file_rosetta
This is needed to automatically boot ppc64_guest_defconfig +
CONFIG_PPC64_BIG_ENDIAN_ELF_ABI_V2=y.

Add both the pie and non-pie variants, even though it is unlikely that
CONFIG_RELOCATABLE will be disabled, as it is selected by a few
different options in the defconfigs that this script will typically be
booting, but it does not hurt to be proactive.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>