buildroot: Rewrite build script in Python
This is the last shell script in this repository and further
improvements to it are needed. Rewrite it in Python to make future
modifications easier, as the majority of ClangBuiltLinux repositories
use Python as the main scripting language now. Remove the shell linting,
as this repository no longer uses it.

As part of the rewrite, this prepares for moving to GitHub releases for
rootfs uploads by moving the final images to 'buildroot/out' versus
'images' directly.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>