boot-utils: Download rootfs images from GitHub releases
Now that rootfs images will be uploaded to GitHub releases, both
boot-qemu.py and boot-uml.py need to download the images before running.

The latest rootfs release can be queried from GitHub's API, which
returns JSON with information about the release tag and assets to allow
easy downloading. Unfortunately, GitHub's API has a rate limit, so
always querying is not possible. To account for this, the script will
first query the rate_limit endpoint to see how many queries are
remaining for the current user. If there are queries remaining, the
script will download the rootfs image if it does not already exist or if
the tag in the '.release' file is different from the latest on. If
there are no queries remaining and the image does not already exist, the
script errors and instructs the user how to make an authenticated API
request or manually download the images themselves from the web.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>