boot-utils: Add '--gh-json-file'
In moving the rootfs images to GitHub releases, we risk hitting GitHub's
API rate limit with GITHUB_TOKEN, which is 1,000 requests per hour per
repository, because each boot test within a workflow will be a separate
call. It is totally possible for us to run 1,000 boots an hour during a
busy workflow period, so this needs special consideration.

To make it easier for CI to cache the results of a GitHub release API
query, add '--gh-json-file' to both boot-qemu.py and boot-uml.py to
allow the tuxsuite parent job to generate boot-utils.json and pass that
along to each child job, so that at worst, each workflow will query the
API three times (once for defconfigs, allconfigs, and distro configs).

Signed-off-by: Nathan Chancellor <nathan@kernel.org>