boot-qemu.py: Handle dtb location shuffle in linux-next
A refactoring commit adjusted the structure of the dts folder in
arch/arm to match arch/arm64 (vendor sub-directories).

Adjust the _dtb member of QEMURunner to be a list of potential locations
for the dtbs, so that they can be found before and after this move,
which works for all kernel versions that this script supports.

Link: https://git.kernel.org/next/linux-next/c/724ba6751532055db75992fc6ae21c3e322e94a7
Signed-off-by: Nathan Chancellor <nathan@kernel.org>