buildroot: Add support for applying local patches
LoongArch support is only available as an in-review patch on the mailing
list. Add scaffolding to make it easy to apply patches on top of the
Buildroot tarball.

To make it easy to add and remove patches while ensuring they get
applied properly, keep the downloaded tarball instead of removing it
after extraction, so that the source can be cleanly recreated each run.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>