boot-qemu.py: Add support for LoongArch
LoongArch requires EFI firmware to boot, which is available as a
precompiled blob from Loongson's GitHub, so it is downloaded when
running 'boot-qemu.py -a loongarch' for the first time. There is no
update mechanism like for the rootfs images for a simple implementation.
If there comes a time where the firmware has to be updated for the
machine to boot properly, such a mechanism can be implemented then.

Other than that, the rest of the support is pretty straightforward. The
defaults for cores and memory are a little higher than other machines,
as that is what other folks are testing.

Closes: https://github.com/ClangBuiltLinux/boot-utils/issues/108
Signed-off-by: Nathan Chancellor <nathan@kernel.org>