boot-qemu.py: Handle difference in arguments between C and Rust implementations of virtiofsd
In at least virtiofsd 1.6.1 (the Rust implementation), the '-o' options
warn that they are deprecated and the '--help' text agrees:

  [2023-07-19T19:32:50Z WARN  virtiofsd] Use of deprecated option format '-o': Please specify options without it (e.g., '--cache auto' instead of '-o cache=auto')

  -o <compat-options>...
          Options in a format compatible with the legacy implementation [deprecated]

To defend against a release removing the deprecated option and breaking
the invocation, maintain two sets of arguments depending on what
implementation is being used. This allows us to drop support for the C
implementation once the Rust one is more widely available in
distributions.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>