utils.py: Do not query GitHub's API at all with '--gh-json-file'
Occasionally, I have noticed the occasional failure to query GitHub's
rate limit API for various reasons, even when providing an up to date
release JSON file via '--gh-json-file'. A rate limit query is unneeded
in this instance because we are not going to query GitHub's API for the
latest release information, as it has already been provided to us by the
user.

Restructure the code so that the rate limit API is only queried when we
will actually need to query the API for the latest release.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>