boot-qemu.py: Allow memory value to be customized
The default memory value for most machines is 512 MB, which is generally
sufficient for most use cases, as these machines are usually quick to
start up and shutdown. Unfortunately, there are times where the machine
may run out of memory when percpu memory allocations are involved along
with a large number of cores, which can happen when booting under KVM
due to the script's heuristics for selecting a good default for '-smp',
which depends on the number of cores on the host machine.

Rather than try to adjust the existing heuristics further, allow users
to specify a memory value that may be better suited for their machine,
just like '-smp'.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>