buildroot/rebuild.py: Add '--savedefconfig'
It is helpful to run savedefconfig when upgrading Buildroot to make sure
that options do not get silently disabled or changed but up until this
point, savedefconfig has only been run when editing a configuration
file, which is annoying when building all images. Add a flag to allow
running savedefconfig without editing so that it can be done
non-interactively.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>